feat: Add RAG project with threaded message retrieval functionality
